requirements.txt
celery==5.6.2
django==6.0.1
django-cors-headers==4.9.0
django-redis==6.0.0
django-rest-framework==0.1.0
djangorestframework==3.16.1
djangorestframework-simplejwt
python-dotenv==1.2.1
redis==7.1.0
python-decouple
pytest
pytest-django
pytest-cov
requests
